pyproject.toml
[project]
name = "system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "python-dateutil>=2.9.0.post0",
    "python-stdnum>=2.2",
]
